Исправляет взаимодействия с сервером
